Integrated metrics AdminServlet.
- Added metrics-servlets dependency to build.gradle

- Added AdminServlet registration to the
  RhizomeWebApplicationInitializer

- Added an important todo to the metrics pod.

Currently, global stats are reported to graphite from every single
instance. A simple way to solve this is to use a Hazelcast distributed
lock to decide which instances is responsible for pushing aggregated
metrics to graphite. If the reporter ever goes down the next lock owner
will start automatically pushing the information.